+++ example/lib/main.dart
@@ -6,7 +6,7 @@ import 'package:sura_flutter_example/examples/stream_future_example.dart';
 import 'examples/buttons_example.dart';
 import 'examples/dialog_popup_example.dart';
 import 'examples/mixin_example.dart';
-import 'examples/sura_accordian_and_sura_expandable.dart';
+import 'examples/sura_accordion_and_sura_expandable.dart';
 import 'examples/tab_indicator_decoration.dart';
 import 'examples/util_example.dart';
 
@@ -18,6 +18,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return SuraProvider(
+      ellipsisText: "No data",
       child: MaterialApp(
         title: 'SuraFlutter Example',
         navigatorKey: SuraNavigator.navigatorKey,
@@ -56,7 +57,7 @@ class _MyHomePageState extends State<MyHomePage> with SuraFormMixin {
       ),
       ExampleButton(
         name: "Sura Accordion and Expandable",
-        child: const SuraAccordianAndExpandable(),
+        child: const SuraAccordionAndExpandable(),
       ),
       ExampleButton(
         name: "Sura Utils",