+++ lib/src/widgets/sura_expandable.dart
@@ -1,26 +1,44 @@
 import 'package:flutter/material.dart';
 
-///A widget that provide that can hide bottomContent
 class SuraExpandable extends StatefulWidget {
+  ///
   final Widget topChild;
+
+  ///
   final Widget bottomChild;
+
+  ///
   final Duration duration;
+
+  ///
   final EdgeInsets padding;
+
+  ///
   final EdgeInsets margin;
+
+  ///Animation's curve
   final Curve curve;
-  final bool initiallyExpand;
-  final Function(bool)? onToggle;
+
+  ///
+  final bool value;
+
+  ///
+  final ValueChanged<bool> onToggle;
+
+  ///
   final bool animatedOnStart;
+
+  ///A widget that provide that can hide bottomContent
   const SuraExpandable({
     Key? key,
     required this.topChild,
     required this.bottomChild,
-    this.initiallyExpand = false,
+    required this.value,
+    required this.onToggle,
     this.duration = const Duration(milliseconds: 200),
     this.padding = const EdgeInsets.all(16),
     this.curve = Curves.linear,
     this.margin = EdgeInsets.zero,
-    this.onToggle,
     this.animatedOnStart = false,
   }) : super(key: key);
   @override
@@ -32,15 +50,12 @@ class _SuraExpandableState extends State<SuraExpandable>
   late AnimationController controller;
   late Animation<double> size;
 
-  void toggle() {
-    if (controller.isAnimating) {
-      return;
-    } else if (controller.isCompleted) {
-      widget.onToggle?.call(false);
-      controller.reverse();
-    } else {
-      widget.onToggle?.call(true);
+  void onToggleAnimation() {
+    if (controller.isAnimating) return;
+    if (widget.value) {
       controller.forward();
+    } else {
+      controller.reverse();
     }
   }
 
@@ -50,7 +65,7 @@ class _SuraExpandableState extends State<SuraExpandable>
       vsync: this,
       duration: widget.duration,
     );
-    if (widget.initiallyExpand) {
+    if (widget.value) {
       widget.animatedOnStart ? controller.forward() : controller.value = 1.0;
     }
     size = CurvedAnimation(curve: widget.curve, parent: controller);
@@ -58,15 +73,17 @@ class _SuraExpandableState extends State<SuraExpandable>
   }
 
   @override
-  void dispose() {
-    controller.dispose();
-    super.dispose();
+  void didUpdateWidget(SuraExpandable oldWidget) {
+    if (widget.value != oldWidget.value) {
+      onToggleAnimation();
+    }
+    super.didUpdateWidget(oldWidget);
   }
 
   @override
-  void didUpdateWidget(SuraExpandable oldWidget) {
-    if (widget.initiallyExpand != oldWidget.initiallyExpand) toggle();
-    super.didUpdateWidget(oldWidget);
+  void dispose() {
+    controller.dispose();
+    super.dispose();
   }
 
   @override
@@ -79,7 +96,7 @@ class _SuraExpandableState extends State<SuraExpandable>
         children: <Widget>[
           InkWell(
             child: widget.topChild,
-            onTap: toggle,
+            onTap: () => widget.onToggle(!widget.value),
           ),
           SizeTransition(
             child: widget.bottomChild,