+++ lib/src/widgets/sura_icon_button.dart
@@ -1,14 +1,31 @@
 import 'package:flutter/material.dart';
 
 class SuraIconButton extends StatelessWidget {
+  ///
   final VoidCallback? onTap;
+
+  ///
   final Widget icon;
+
+  ///
   final EdgeInsets margin;
+
+  ///
   final EdgeInsets padding;
+
+  ///
   final Color backgroundColor;
+
+  ///
   final double borderRadius;
+
+  ///
   final double elevation;
+
+  ///
   final BorderSide? borderSide;
+
+  ///
   final Widget? badge;
 
   ///An IconButton with respectively small margin and shape
@@ -52,7 +69,7 @@ class SuraIconButton extends StatelessWidget {
       child: InkWell(
         onTap: onTap,
         customBorder: shape,
-        mouseCursor: MouseCursor.defer,
+        mouseCursor: SystemMouseCursors.click,
         child: Stack(
           children: [
             Padding(