+++ lib/src/widgets/sura_raised_button.dart
@@ -1,35 +1,66 @@
+// ignore_for_file: must_be_immutable
+
 import 'package:flutter/material.dart';
 
 import 'conditional_widget.dart';
 import 'spacing.dart';
 import 'sura_provider.dart';
 
-/// SuraRaisedButton can be use sometime to replace RaisedButton or ElevatedButton because we provide more flexibility and wrap around
 class SuraRaisedButton extends StatelessWidget {
   ///receive a ValueNotifier to indicate a loading widget
   final ValueNotifier<bool>? loadingNotifier;
+
+  ///
   final Widget child;
+
+  ///An icon to show at before [child]
   final Widget? icon;
+
+  ///
   final VoidCallback? onPressed;
+
+  ///
   final Function? onLongPressed;
+
+  //
   final double? elevation;
-  //Button's background Color
+
+  ///Button's background Color
   final Color? color;
+
+  ///Text's color for a child that usually a Text
   final Color? textColor;
-  //Loading indicator's color
+
+  ///Loading indicator's color, default is [white]
   final Color loadingColor;
+
+  ///A widget to show when loading, if the value is null,
+  ///it will use a loading widget from SuraProvider or CircularProgressIndicator
   final Widget? loadingWidget;
+
+  ///Button's margin
   final EdgeInsets margin;
+
+  ///Button's padding
   final EdgeInsets padding;
+
+  ///Button's shape, default is [StadiumBorder]
   final OutlinedBorder? shape;
+
+  ///child's alignment
   final MainAxisAlignment? alignment;
 
   ///if [fullWidth] is `true`, Button will take all remaining horizontal space
   final bool fullWidth;
+
+  ///
   final BorderSide? borderSide;
 
+  ///
+  final GlobalKey _globalKey = GlobalKey();
+
   ///Create a button with loading notifier
-  const SuraRaisedButton({
+  SuraRaisedButton({
     Key? key,
     required this.onPressed,
     required this.child,
@@ -48,6 +79,19 @@ class SuraRaisedButton extends StatelessWidget {
     this.alignment,
     this.borderSide,
   }) : super(key: key);
+
+  void maintainWidthOnLoading() {
+    if (fullWidth == false && width == null) {
+      WidgetsBinding.instance?.addPostFrameCallback((d) {
+        RenderBox box =
+            _globalKey.currentContext!.findRenderObject() as RenderBox;
+        width = box.size.width;
+      });
+    }
+  }
+
+  double? width;
+
   @override
   Widget build(BuildContext context) {
     return Container(
@@ -57,6 +101,7 @@ class SuraRaisedButton extends StatelessWidget {
       child: ValueListenableBuilder<bool>(
         valueListenable: loadingNotifier ?? ValueNotifier(false),
         builder: (context, loading, _) {
+          maintainWidthOnLoading();
           return ElevatedButton(
             onPressed: loading ? () {} : onPressed,
             style: ElevatedButton.styleFrom(
@@ -71,6 +116,7 @@ class SuraRaisedButton extends StatelessWidget {
             child: ConditionalWidget(
               condition: loading,
               onFalse: () => Row(
+                key: _globalKey,
                 mainAxisAlignment: alignment ?? MainAxisAlignment.center,
                 mainAxisSize: MainAxisSize.min,
                 children: <Widget>[
@@ -94,10 +140,15 @@ class SuraRaisedButton extends StatelessWidget {
 
   Widget _buildLoadingWidget() {
     return SizedBox(
-      width: icon != null ? 24 : 20,
-      height: icon != null ? 24 : 20,
-      child: CircularProgressIndicator(
-        valueColor: AlwaysStoppedAnimation<Color>(loadingColor),
+      width: width,
+      child: Center(
+        child: SizedBox(
+          width: icon != null ? 24 : 20,
+          height: icon != null ? 24 : 20,
+          child: CircularProgressIndicator(
+            valueColor: AlwaysStoppedAnimation<Color>(loadingColor),
+          ),
+        ),
       ),
     );
   }